Fixed back edge detection to work with loops with multiple entries.
The back edge in a loop with multiple entries is ambiguous and will appear different depending on which entry you pick. findBackEdgeDestinations now arbitrarily picks one instead of picking both and screwing up the single-entry tranformation.

That's an interesting caveat in the No More Gotos paper. I don't recall the authors addressing this at all.