SESELoop: added code to turn un-dominating instructions into PHI nodes
This solves a problem with the SESELoop pass where an instruction from a loop
would be used in loop successors after a single-exit node was inserted. This
would cause that instruction to no longer dominate all its uses. This problem
is solved by inserting PHI nodes in the funnel block.