Fixed issue with loop conditions not being considered uses
This adds an ExpressionNode object to every LoopNode and IfElseNode. This
expression node is statically allocated into the loop/if-else. The condition
pointer is now a reference to that embedded expression.

This approach allows the variable references pass to dedicate a statement index
to the condition itself. Otherwise, the statement index refers to the control
structure, and it makes it look like post-tested loops have a pre-tested
condition.