Created pass to detect and simplify sign-ext pattern
The new SignExt pass detects this IR pattern:

  %1 = /* i32 */
  %2 = ashr i32 %1, 31
  %3 = zext i32 %2 to i64
  %4 = shl nuw i64 %3, 32
  %5 = zext i32 %1 to i64
  %6 = or i64 %4, %5

and transforms it into a much simpler, single sext instruction:

  %1 = /* i32 */
  %2 = sext i32 %1 to i64

It correctly handles non-int32 cases as well, but int32 to in64 is the most
common and therefore is used as an example.

Additionally, this commit annotates output cast expressions with the `__sext` or
`__zext` pseudo-C keyword to clarify whether the cast should sign-extend or
zero-extend its target, since we don't have working type inference and things
quickly get confusing otherwise.