Split register use registry from register use analysis
This commit creates a new IpaRegisterUse class that performs interprocedural
register use analysis. This analysis needs to operate on a whole module and is
cumbersome in situations where you only want to disassemble a small part of
an executable.

The RegisterUse pass is then used as an analysis-independent argument registry.
The separation makes it possible to have multiple independent argument discovery
passes, which is a useful step forward to solve the imported symbol hack *and*
the partial disassembly case.