Create ExpressionContext class
This is the first step of a wider effort to change how expressions work. Instead
of having one statement per instruction and then try to simplify that,
expressions will have a use list and tools to work with them.

The end goal is to implement more powerful analyses and get rid of the
expression visiting mechanism that we've now significantly outgrown.