Stop using call graph only to update it
This was never really understood because it was cargo culted off LLVM's argument
promotion pass, but you only need to update the call graph if you use it within
the pass. Fcd probably used it the wrong way somewhere and it crashed with the
stable LLVM 3.8 release, so this commit does away with it entirely.