Work around ABI compatibility between Clang and GCC
Debian's llvm-3.8 package was compiled with the "C++11 ABI", which Clang is unaware of at this point. Any function that has std::list or std::string as a parameter or return value compiled from GCC cannot be linked to using Clang. The only instance in fcd was llvm::sys::Process::GetEnv, so it's been replaced with a mere gentev call.