Create ExpressionReference "smart expression pointer"
In a world where expressions and statements gain a `dropAllReferences` method that applies recursively to expressions as they are dropped, we want a type that will keep expressions alive while the AST is being manipulated. The ExpressionReference type has an explicit version of the use layout that is practical to have as a local variable.