Stop always creating a result variable for calls
Calls previously always created a result variable to avoid ordering problems. This is because calls, like loads and stores, are memory operations with side effects and are therefore ordering-sensitive. With this commit, fcd can eliminate the result variable for a call if the following is true:

1) the call has just one use other than its declaration
2) there are no memory operations between the call and its use

This helps simplify code like this:

	uint16_t anon11 = htons((uint16_t)alloca1.field9);
	alloca1.field16 = anon11;

into a simpler `alloca1.field16 = htons((uint16_t)alloca1.field9)`.